DP-1338 add protractor test that we asserts that we can add a user that already exists in another tenant.
Changing the functional test setup by adding a new user in only one tenant.
Then in the protractor test, add the user to the other tenant and assert that it has been added. 
Also, contains refactor functional tests such that all userService/GlobalAuth actions are done in le-security,
and we use the new InternalTestUserService.