DCP-1083
In EmailServiceImpl a second EmailSettings bean that holds the SendGrid configuration is added and a new method to provide the desired EmailSettings bean to the send email methods.

In EmailUtils changed to use the Transport.send() method that takes the email provider username and password.  Also, marked an internal method as private.

Added a comment about a unused method to GlobalUserManagementServiceImpl.

For all the monitor.properties files, added the config information for SendGrid.

In monitor-email-settings-context.xml, added the new settings for the second EmailSettings bean.