DCP-1083
In le-parent/pom.xml added an exclusion to exclude JavaMail.  The JavaMail that this dependency was dependent on is an old version and due to the JavaMail project changing it's groupId and artifactId values maven could not automatically find the new version.

Added a comment to monitor-email-settings-context.xml to make clear that apikey is the username for the sendgrid account.

Updated and refactored EmailService and EmailServiceImpl to make it simplier to specify the From address and therefore which email service provider to use when sending each email message.

Added some additional tests to EmailServiceImplTestNG.