DCP-1861
In LedpCode, added two new messages for the case where a project is being created but the Purpose Of Use is not entitled.

In ProjectServiceImpl added new methods to validate the Tenant has the correct entitlements for the given Purpose of Use.

In DCPFunctionalTestNGBase updated the getPurposeOfUse method so it is a Purpose of Use that the Tenant is entitled to.

In ProjectServiceImplTestNG created tests for testing non-entitled Purpose Of Use.